requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-docx>=1.1.0
beautifulsoup4
lxml
openpyxl
pdfplumber
python-docx
requests
